pyproject.toml
[project]
name = "test-bench"
version = "0.1.0"
description = "Testbench to evaluate agents using Ragas"
requires-python = ">=3.12"
dependencies = [
    "a2a>=0.44",
    "a2a-sdk>=0.3.10",
    "httpx>=0.28.1",
    "langchain-openai>=1.0.2",
    "pandas>=2.3.3",
    "pandas-stubs>=2.3.0",
    "pyarrow>=21.0.0",
    "ragas>=0.3.5",
    "requests>=2.32.5",
    "types-requests>=2.32.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp-proto-http>=1.20.0",
    "opentelemetry-instrumentation-httpx>=0.45b0",
]

[dependency-groups]
dev = [
    "bandit[toml]>=1.7.8",
    "mypy>=1.17.0",
    "import-linter>=2.0",
    "poethepoet>=0.31.1",
    "ruff>=0.12.7",
    "pytest>=7.4.0",
    "pytest-asyncio>=1.3.0",
]

[tool.poe.tasks]
test = "pytest tests/ -v"
test_e2e = "pytest tests_e2e/ -v"

format = "ruff format"
lint = "ruff check --fix"
ruff = ["format", "lint"]

bandit = "bandit -c pyproject.toml -r ."
mypy = "mypy ."

check = ["test", "mypy", "bandit", "ruff"]

[tool.ruff]
# Allow lines to be as long as 120.
line-length = 120


[tool.ruff.lint]
# https://docs.astral.sh/ruff/rules/
# I001 --> enable import sorting rules
# Q --> enable flake8-quotes rules for consistent quote style
# N --> pep8-naming for proper naming conventions
# PT --> flake8-pytest-style for pytest best practices
# TID251 --> banned relative imports
# DTZ --> datetime timezone awareness
extend-select = ["Q", "N", "I001", "PT", "TID251", "DTZ"]

[tool.mypy]
mypy_path = "scripts"

[tool.bandit]
exclude_dirs = [".cache", ".venv"]

[tool.bandit.assert_used]
skips = ["*_test.py", "*test_*.py"]
